Clarify description of Flux functionality
Clarifies the description of Flux functionality in the README. The following text on lines 41 - 43 is difficult to understand:

BEFORE: "When a user interacts with a React view, the view propagates an action through a central dispatcher, to the various stores that hold the application's data and business logic, which updates all of the views that are affected."

AFTER: "When a user interacts with a React view, the view propagates an action through a central dispatcher to the various stores that hold the application's data and business logic, updating all affected views."

In the original case, the "which" in "which updates all of the views that are affected" refers to an ambiguous subject (could be the React view, the central dispatcher, the stores or the data and business logic). I would like to change this sentence so that the gerund implies that propagating an action from the view down to the stores is causing all affected views to be updated.

I had some trouble initially understanding what this sentence was trying to say and I think my change might relieve some possible confusion. I hope this is helpful - I was trying to explain Flux to my class and we spent some time parsing out the README.