Update TodoList doc to reflect the implementation
The code for the todomvc example has changed but the documentation was
not updated. Many times the doc invites the reader to read the full
source, so a sync between doc and code might be required now.

In particular, the dispatcher is handwritten and makes use of promises,
which is neat but can lead to much confusion when reading the source as
a reference, especially since this Todo example acts as an introduction
to the Flux architecture.

I hereby propose to reduce the complexity of the documentation by
sticking to what's in the source.

Signed-off-by: Jean-Denis Vauguet <jd@vauguet.fr>