fixed typo
Thread object has no `lastTimestamp` property, hence the condition is not valid. The main task of this iteration is to keep an object of the data of the last message, so the fix is to have the comparison condition between the `timestamp`s of the last kept message and the current one from the iteration.