Importing Counter and Todo files.
I'm not sure if the intentions were for the user to have the option of creating their own Counter file, but currently  it is unclear where the `Counter` and `Todo` files are being imported. I edited lines 241 and 245 to clarify so the user knows to import these two files. I welcome feedback .