Improve performance of FluxContainerSubscriptions
A `FluxContainer` recomputes the stores list (the one created by
`getStores`) after every change in `props` and `context`. It is needed
because a container can use `props` to refine how the `state` is
computed. The `FluxContainerSubscriptions` manages when a container
should be recomputed. It is done listening for changes in dependent
stores. In order to do that, `FluxContainerSubscriptions` creates a
store group that is a fake store that wait for all dependent stores.
If the stores change during the dispatching cycle, then it call the
listener which one recompute the state of the container.

This PR changes `FluxContainerSubscriptions` to only create new
listener and a new store group when the stores list changes.