Early bail from setState in componentWillReceiveProps
Right now every single time the container receive props and has the withProps or withContext options we are going to schedule an extra render by calling setState, this makes render and getState to overfire. Both of this functions are blockers for frame rendering. 

In order to make this better we can add a shallow equal comparison for props and context. because getStores and getState are pure functions we can assume if we pass the same parameters we are going to get the same data.