Fix incomplete code example in README.md (#458)
* Fix incomplete code example in README.md

In step 4, the guide says "update AppView.js" but does not provide `...` to show where the new code should be inserted. Because of this, the user is likely to omit the crucial statements `import React from 'react';` and `export default AppView;`. Simply providing these statements is the easiest solution.

* Fix more incomplete code examples in README.md

Provide `...` and explicit imports for remaining examples in the README.md.