Update URLs in yarn.lock to fix yarn install
Without this patch, yarn install will fail due to
artifactory.twitter.biz not resolving. To fix this I deleted all of the
"resolved" and "integrity" lines out of the yarn.lock file and ran "yarn
install" to pull in updated URLs.

A few version numbers also changed and the dist files were altered, but
the test suite still passes and it doesn't seem that anything
functionally changed.